+++ engine/src/engine/ui/systems/ScrollViewUpdateSystem.cpp
@@ -67,7 +67,7 @@ namespace se::ui::systems
                                 auto oldTransform = childTransform;
                                 childTransform.anchors = { 0, 1, 0, 0 };
                                 auto desiredSize = DesiredSizeCalculator::GetDesiredSize(this, child, &widgets[i], rectTransform, childTransform, &value[i]);
-                                childTransform.maxY = static_cast<int>(childTransform.minY + desiredSize.y / window->GetContentScale());
+                                childTransform.maxY = childTransform.minY + desiredSize.y / window->GetContentScale();
                                 childTransform.rect = util::CalculateScreenSpaceRect(childTransform, rectTransform);
                                 totalChildSize += childTransform.rect.size.y;
                                 if (!childTransform.overridesChildSizes)
@@ -88,7 +88,7 @@ namespace se::ui::systems
                 float availableScrollSpaceBottom = std::max(0.f, totalChildSize - rectTransform.rect.size.y);
                 if (availableScrollSpaceBottom > 0.f)
                 {
-                    util::TranslateChildren(entity, this, math::IntVec2(0, -availableScrollSpaceBottom * scrollView.scrollAmount));
+                    util::TranslateChildren(entity, this, math::IntVec2(0, static_cast<int>(-availableScrollSpaceBottom * scrollView.scrollAmount)));
                     scrollView.onScrolled.Broadcast(&rectTransform, scrollView.scrollAmount);
                 }
